Change all fortran real to real*8

+++ src/fortran_routing/mc_pylink_v00/MC_singleSeg_singleTS/varSingleSegStime_f2py_noIC_clean.f90
@@ -2,8 +2,8 @@ module var
     implicit none
     save
 
-    real :: dt, dx, qup, quc, qdp, qdc, ql, Bw, Tw, TwCC, nCC, Cs, So, n, z, vel, depth
-    real :: bfd, WPC, AREAC, C1, C2, C3, C4
+    real*8 :: dt, dx, qup, quc, qdp, qdc, ql, Bw, Tw, TwCC, nCC, Cs, So, n, z, vel, depth
+    real*8 :: bfd, WPC, AREAC, C1, C2, C3, C4
 
 end module var
